pyproject.toml
[tool.poetry]
name = "restapi-app"
version = "0.1.0"
description = ""
authors = ["Nazar Chaban <134311145+Vivern0@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.1"
sqlalchemy = "^2.0.29"
uvicorn = "^0.29.0"
alembic = "^1.13.1"
psycopg2 = "^2.9.9"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.9"
bcrypt = "^4.1.2"
pydantic-settings = "^2.2.1"
fastapi-mail = "^1.4.1"
cloudinary = "^1.40.0"
redis = "^5.0.4"
fastapi-limiter = "^0.1.6"
httpx = "^0.27.0"


[tool.poetry.group.dev.dependencies]
sphinx = "^7.3.7"
pytest = "^8.2.0"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
